Get first form association example working
We now have a set of check boxes working for administrators to students.
The next step will be to set up an association that is a little deeper.
Something like a student supply list that is also a part of a student
admission form. I think that will generate the level of complexity I'm
attempting to sort out.